database_schema_agent/requirements.txt
pika
python-dotenv
openai
